removed redundant link, added archive.org links
The link to "Sqrrl Hunting Techniques" was the same as the "Sqrrl Guide to Threat Hunting", so I changed the link to the "huntpedia" by Sqrrl. In addition I changed the links to the pdf files in the resource section to web.archive.org links, as the future of these on sqrrl.com is uncertain.